Changed location of Question mark icon to avoid up page arrow from overlapping into the youtube's component
